// apps/a11y-tests/src/axe-sarif-converter/sarif/sarifv2.d.ts
export interface File {
    /**
     * The location of the file.
     */
    fileLocation?: {
        /**
         * A string containing a valid relative or absolute URI.
         */
        uri: string;
        /**
         * A string which indirectly specifies the absolute URI with respect to which a relative URI in the "uri" property is interpreted.
         */
        uriBaseId?: string;
    };
    /**
     * Identifies the key of the immediate parent of the file, if this file is nested.
     */
    parentKey?: string;
    /**
     * The offset in bytes of the file within its containing file.
     */
    offset?: number;
    /**
     * The length of the file in bytes.
     */
    length?: number;
    /**
     * The role or roles played by the file in the analysis.
     */
    roles?: (
        | 'analysisTarget'
        | 'attachment'
        | 'responseFile'
        | 'resultFile'
        | 'standardStream'
        | 'traceFile'
        | 'unmodifiedFile'
        | 'modifiedFile'
        | 'addedFile'
        | 'deletedFile'
        | 'renamedFile'
        | 'uncontrolledFile')[];
    /**
     * The MIME type (RFC 2045) of the file.
     */
    mimeType?: string;
    /**
     * The contents of the file.
     */
    contents?: {
        /**
         * UTF-8-encoded content from a text file.
         */
        text?: string;
        /**
         * MIME Base64-encoded content from a binary file, or from a text file in its original encoding.
         */
        binary?: string;
    };
    /**
     * Specifies the encoding for a file object that refers to a text file.
     */
    encoding?: string;
    /**
     * An array of hash objects, each of which specifies a hashed value for the file, along with the name of the hash function used to compute the hash.
     */
    hashes?: Hash[];
    /**
     * The date and time at which the file was most recently modified. See "Date/time properties" in the SARIF spec for the required format.
     */
    lastModifiedTime?: string;
    /**
     * Key/value pairs that provide additional information about the file.
     */
    properties?: {
        /**
         * A set of distinct strings that provide additional information.
         */
        tags?: string[];
        [k: string]: any;
    };
}
/**
 * A hash value of some file or collection of files, together with the hash function used to compute the hash.
 */
export interface Hash {
    /**
     * The hash value of some file or collection of files, computed by the hash function named in the 'algorithm' property.
     */
    value: string;
    /**
     * The name of the hash function used to compute the hash value specified in the 'value' property.
     */
    algorithm: string;
}
// ...
export interface Result {
    /**
     * The stable, unique identifier of the rule (if any) to which this notification is relevant. This member can be used to retrieve rule metadata from the rules dictionary, if it exists.
     */
    ruleId?: string;
    /**
     * A value specifying the severity level of the result.
     */
    level?: 'notApplicable' | 'pass' | 'note' | 'warning' | 'error' | 'open';
    /**
     * A message that describes the result. The first sentence of the message only will be displayed when visible space is limited.
     */
    message?: {
        /**
         * A plain text message string.
         */
        text?: string;
        /**
         * The resource id for a plain text message string.
         */
        messageId?: string;
        /**
         * A rich text message string.
         */
        richText?: string;
        /**
         * The resource id for a rich text message string.
         */
        richMessageId?: string;
        /**
         * An array of strings to substitute into the message string.
         */
        arguments?: string[];
    };
    /**
     * Identifies the file that the analysis tool was instructed to scan. This need not be the same as the file where the result actually occurred.
     */
    analysisTarget?: {
        /**
         * A string containing a valid relative or absolute URI.
         */
        uri: string;
        /**
         * A string which indirectly specifies the absolute URI with respect to which a relative URI in the "uri" property is interpreted.
         */
        uriBaseId?: string;
    };
    /**
     * One or more locations where the result occurred. Specify only one location unless the problem indicated by the result can only be corrected by making a change at every specified location.
     */
    locations?: Location[];
    /**
     * A stable, unique identifer for the result in the form of a GUID.
     */
    instanceGuid?: string;
    /**
     * A stable, unique identifier for the equivalence class of logically identical results to which this result belongs, in the form of a GUID.
     */
    correlationGuid?: string;
    /**
     * A set of strings that contribute to the stable, unique identity of the result.
     */
    partialFingerprints?: {
        [k: string]: string;
    };
    /**
     * A set of strings each of which individually defines a stable, unique identity for the result.
     */
    fingerprints?: {
        [k: string]: string;
    };
    /**
     * An array of 'stack' objects relevant to the result.
     */
    stacks?: Stack[];
    /**
     * An array of 'codeFlow' objects relevant to the result.
     */
    codeFlows?: CodeFlow[];
    /**
     * An array of one or more unique 'graph' objects.
     */
    graphs?: Graph[];
    /**
     * An array of one or more unique 'graphTraversal' objects.
     */
    graphTraversals?: GraphTraversal[];
    /**
     * A set of locations relevant to this result.
     */
    relatedLocations?: Location[];
    /**
     * A set of flags indicating one or more suppression conditions.
     */
    suppressionStates?: ('suppressedInSource' | 'suppressedExternally')[];
    /**
     * The state of a result relative to a baseline of a previous run.
     */
    baselineState?: 'new' | 'existing' | 'absent';
    /**
     * A set of files relevant to the result.
     */
    attachments?: Attachment[];
    /**
     * The URIs of the work items associated with this result
     */
    workItemUris?: string[];
    /**
     * An array of physicalLocation objects which specify the portions of an analysis tool's output that a converter transformed into the result object.
     */
    conversionProvenance?: PhysicalLocation[];
    /**
     * An array of 'fix' objects, each of which represents a proposed fix to the problem indicated by the result.
     */
    fixes?: Fix[];
    /**
     * Key/value pairs that provide additional information about the result.
     */
    properties?: {
        /**
         * A set of distinct strings that provide additional information.
         */
        tags?: string[];
        [k: string]: any;
    };
}
// ...
export interface Rule {
    /**
     * A stable, opaque identifier for the rule.
     */
    id: string;
    /**
     * A rule identifier that is understandable to an end user.
     */
    name?: {
        /**
         * A plain text message string.
         */
        text?: string;
        /**
         * The resource id for a plain text message string.
         */
        messageId?: string;
        /**
         * A rich text message string.
         */
        richText?: string;
        /**
         * The resource id for a rich text message string.
         */
        richMessageId?: string;
        /**
         * An array of strings to substitute into the message string.
         */
        arguments?: string[];
    };
    /**
     * A concise description of the rule. Should be a single sentence that is understandable when visible space is limited to a single line of text.
     */
    shortDescription?: {
        /**
         * A plain text message string.
         */
        text?: string;
        /**
         * The resource id for a plain text message string.
         */
        messageId?: string;
        /**
         * A rich text message string.
         */
        richText?: string;
        /**
         * The resource id for a rich text message string.
         */
        richMessageId?: string;
        /**
         * An array of strings to substitute into the message string.
         */
        arguments?: string[];
    };
    /**
     * A description of the rule. Should, as far as possible, provide details sufficient to enable resolution of any problem indicated by the result.
     */
    fullDescription?: {
        /**
         * A plain text message string.
         */
        text?: string;
        /**
         * The resource id for a plain text message string.
         */
        messageId?: string;
        /**
         * A rich text message string.
         */
        richText?: string;
        /**
         * The resource id for a rich text message string.
         */
        richMessageId?: string;
        /**
         * An array of strings to substitute into the message string.
         */
        arguments?: string[];
    };
    /**
     * A set of name/value pairs with arbitrary names. The value within each name/value pair consists of plain text interspersed with placeholders, which can be used to construct a message in combination with an arbitrary number of additional string arguments.
     */
    messageStrings?: {
        [k: string]: string;
    };
    /**
     * A set of name/value pairs with arbitrary names. The value within each name/value pair consists of rich text interspersed with placeholders, which can be used to construct a message in combination with an arbitrary number of additional string arguments.
     */
    richMessageStrings?: {
        [k: string]: string;
    };
    /**
     * Information about the rule that can be configured at runtime.
     */
    configuration?: {
        /**
         * Specifies whether the rule will be evaluated during the scan.
         */
        enabled?: boolean;
        /**
         * Specifies the default severity level of the result.
         */
        defaultLevel?: 'note' | 'warning' | 'error' | 'open';
        /**
         * Contains configuration information specific to this rule.
         */
        parameters?: {
            /**
             * A set of distinct strings that provide additional configuration information.
             */
            tags?: string[];
            [k: string]: any;
        };
    };
    /**
     * A URI where the primary documentation for the rule can be found.
     */
    helpUri?: string;
    /**
     * Provides the primary documentation for the rule, useful when there is no online documentation.
     */
    help?: {
        /**
         * A plain text message string.
         */
        text?: string;
        /**
         * The resource id for a plain text message string.
         */
        messageId?: string;
        /**
         * A rich text message string.
         */
        richText?: string;
        /**
         * The resource id for a rich text message string.
         */
        richMessageId?: string;
        /**
         * An array of strings to substitute into the message string.
         */
        arguments?: string[];
    };
    /**
     * Key/value pairs that provide additional information about the rule.
     */
    properties?: {
        /**
         * A set of distinct strings that provide additional information.
         */
        tags?: string[];
        [k: string]: any;
    };
}